14387 track add item analytics in aggregate model
