+++ WooCommerce/Classes/POS/Presentation/CartView.swift
@@ -254,8 +254,7 @@ import class WooFoundation.MockAnalyticsProviderPreview
                                           cardPresentPaymentService: CardPresentPaymentPreviewService(),
                                           currencyFormatter: .init(currencySettings: .init()),
                                           paymentState: .acceptingCard)
-    let cartViewModel = CartViewModel(posModel: PointOfSaleAggregateModel(itemProvider: POSItemProviderPreview()),
-                                      analytics: MockAnalyticsPreview())
+    let cartViewModel = CartViewModel(posModel: PointOfSaleAggregateModel(itemProvider: POSItemProviderPreview()))
     let itemsListViewModel = ItemListViewModel(posModel: PointOfSaleAggregateModel(itemProvider: POSItemProviderPreview()))
     let dashboardViewModel = PointOfSaleDashboardViewModel(posModel: PointOfSaleAggregateModel(itemProvider: POSItemProviderPreview()),
                                                            cardPresentPaymentService: CardPresentPaymentPreviewService(),
