+++ WooCommerce/Classes/POS/Presentation/PointOfSaleEntryPointView.swift
@@ -27,8 +27,7 @@ struct PointOfSaleEntryPointView: View {
                                               cardPresentPaymentService: cardPresentPaymentService,
                                               currencyFormatter: currencyFormatter,
                                               paymentState: .acceptingCard)
-        let cartViewModel = CartViewModel(posModel: posModel,
-                                          analytics: analytics)
+        let cartViewModel = CartViewModel(posModel: posModel)
         let shouldShowGhostableItemCard = ServiceLocator.featureFlagService.isFeatureFlagEnabled(.displayInfiniteScrollingUIDetailsInPointOfSale)
         let itemListViewModel = ItemListViewModel(posModel: posModel, shouldShowGhostableItemCard: shouldShowGhostableItemCard)
